requirements.txt
django
phonenumbers
django-phonenumber-field
django-crispy-forms #bootstrap forms

